% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.menu-package.R
\docType{package}
\name{cli.menu-package}
\alias{cli.menu}
\alias{cli.menu-package}
\title{cli.menu: A multi-select commmand line menu}
\description{
A simple command line menu with more than one options to select.
}
\author{
\strong{Maintainer}: Mark Heckmann \email{heckmann.mark@gmail.com}

}
\keyword{internal}
