% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menu.R
\name{cli_menu}
\alias{cli_menu}
\title{Interactive multi-select menu for command line}
\usage{
cli_menu(
  choices,
  prompt = "Your selection > ",
  all_opt = FALSE,
  none_opt = FALSE
)
}
\arguments{
\item{choices}{A vector with items to choose from.}
}
\description{
Interactive multi-select menu for command line
}
\examples{
if (interactive()) {
  x <- c("One", "Two", "Three")
  cli_menu(x)
}
}
