% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menu.R
\name{collect_integers}
\alias{collect_integers}
\title{Collect integers from string}
\usage{
collect_integers(x)
}
\arguments{
\item{x}{A string.}
}
\description{
Parses a string and extracts all integers including range-like expressions. Integers can be separated by
any character. Integers seperated by \code{-} or \code{:} are interpreted as a range.
}
\examples{

# single integers
x <- "1 2 3"
collect_integers(x)

x <- "1, 2, 3"
collect_integers(x)

# use ranges
x <- "1-3"
collect_integers(x)

x <- "1:3, 5:6"
collect_integers(x)

# mix integers and ranges
x <- "1:3, 4, 7:8"
collect_integers(x)

# Parser handles all kinds of less structured scenarios

x <- " 3, 4, 5-7,10, 11- 14, 20 :22"
collect_integers(x)

x <- "1 2, 4;5 hello world || 6 A11A"
collect_integers(x)

x <- "A1,5;;;6-  8H11:  15"
collect_integers(x)

}
\keyword{internal}
