% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menu.R
\name{get_ranges}
\alias{get_ranges}
\title{Extract ranges}
\usage{
get_ranges(x)
}
\arguments{
\item{x}{A string.}
}
\description{
Integers seperated by \code{-} or \code{:} are interpreted as a range.
}
\keyword{internal}
