% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menu.R
\name{range_to_numeric_vector}
\alias{range_to_numeric_vector}
\title{Convert range string to numeric vector}
\usage{
range_to_numeric_vector(x)
}
\arguments{
\item{x}{A string.}
}
\description{
Convert range string to numeric vector
}
\keyword{internal}
