% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menu.R
\name{remove_ranges}
\alias{remove_ranges}
\title{Remove ranges from string}
\usage{
remove_ranges(x)
}
\arguments{
\item{x}{A string.}
}
\description{
Remove all parts which are interpretetd as a range from string.
}
\keyword{internal}
