% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menu.R
\name{cli_menu}
\alias{cli_menu}
\title{Interactive multi-select menu for command line}
\usage{
cli_menu(
  choices,
  prompt_1 = "Please select from list:",
  prompt_2 = "Your selection > ",
  all_opt = FALSE,
  none_opt = FALSE
)
}
\arguments{
\item{choices}{A vector with items to choose from.}

\item{prompt_1, prompt_2}{User prompts before and after selection list.}

\item{all_opt, none_opt}{Not yet implemented.}
}
\description{
Interactive multi-select menu for command line
}
\examples{
if (interactive()) {
  x <- c("One", "Two", "Three")
  cli_menu(x)
}
}
